CU-85zrx08ta: Integrate Eveara Distribution Service addtl. work

+++ newm-server/src/main/kotlin/io/newm/server/features/distribution/model/EvearaOutlet.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class EvearaOutlet(
     @SerialName("process_duration_dates")
-    val processDurationDates: Int,
+    val processDurationDates: Long,
     @SerialName("dolby_enabled")
     val dolbyEnabled: Boolean,
     @SerialName("store_name")
