+++ newm-server/src/main/kotlin/io/newm/server/features/distribution/model/GetUserResponse.kt
@@ -10,7 +10,7 @@ data class GetUserResponse(
     @SerialName("success")
     val success: Boolean,
     @SerialName("total_records")
-    val totalRecords: Int,
+    val totalRecords: Long,
     @SerialName("data")
     val users: List<UserResponse>,
 )
